+++ www/api/auth.php
@@ -234,8 +234,15 @@ try {
 		
 		//Server side cookie worked better on safari. Client side cookies were removed on reboot.
 		$expires = !empty($data['rememberLogin']) ? strtotime("+1 year") : 0;
-		setcookie('accessToken', $token->accessToken, $expires, "/", Request::get()->getHost(), false, false);
-		
+
+
+		Response::get()->setCookie('accessToken', $token->accessToken, [
+			'expires' => $expires,
+			"path" => "/",
+			"samesite" => "Strict",
+			"domain" => Request::get()->getHost()
+		]);
+
 		output($response, 201, "Authentication is complete, access token created.");
 	}