+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Core: Validate theme and language. Don't echo given language string to fix XSS issue
+- Core: Set "samesite=Strict" on all cookies
 
 12-01-2021 6.4.210
 - Core: Fixed constraint error in database check